perf: move ref sequence transform out of the loop
Unfortunately, this now publicly exposes many previously internal interfaces - types and functions
