+++ packages/nextalign/benchmarks/src/ForwardTrace.benchmark.h
@@ -5,8 +5,8 @@
 #include <numeric>
 #include <vector>
 
-#include "../../src/align/alignPairwise.h"
 #include "../include/nextalign/nextalign.h"
+#include "../src/align/alignPairwise.h"
 #include "utils/getData.h"
 #include "utils/setCounters.h"
