+++ packages/nextalign/benchmarks/src/Nextalign.benchmark.h
@@ -5,18 +5,29 @@
 #include <numeric>
 #include <vector>
 
-#include "../../src/align/alignPairwise.h"
 #include "../include/nextalign/nextalign.h"
+#include "../src/align/alignPairwise.h"
 #include "utils/getData.h"
 #include "utils/setCounters.h"
 
 
-/**
- * Average benchmark for nextalign().
- * Runs `nextalign()` for NUM_SEQUENCES_AVG sequences and averages the result.
- * This is an estimate of runtime performance in a real world scenario, when many sequences are ran in a batch.
- */
-void NextalignAverage(benchmark::State& st) {
+class NextalignAverageBench : public benchmark::Fixture {
+protected:
+  std::vector<NucleotideSequence> nucSequences;
+  NucleotideSequence ref = toNucleotideSequence(reference);
+
+  NextalignAverageBench() {
+    const auto n = NUM_SEQUENCES_AVG;
+    nucSequences.resize(n);
+    for (int i = 0; i < n; ++i) {
+      const auto& input = sequences[i];
+      nucSequences[input.index] = toNucleotideSequence(input.seq);
+    }
+  }
+};
+
+
+BENCHMARK_DEFINE_F(NextalignAverageBench, Average)(benchmark::State& st) {
   const auto n = NUM_SEQUENCES_AVG;
   const NextalignOptions options = {};
   Alignment aln;
@@ -24,41 +35,41 @@ void NextalignAverage(benchmark::State& st) {
 
   for (const auto _ : st) {
     for (int i = 0; i < n; ++i) {
-      const auto& input = sequences[i];
-      benchmark::DoNotOptimize(aln = nextalign(input.seq, reference, geneMap, options));
+      const auto& seq = nucSequences[i];
+      benchmark::DoNotOptimize(aln = nextalign(seq, ref, geneMap, options));
     }
   }
 
   setCounters(st, n);
 }
 
-BENCHMARK(NextalignAverage)      //
+BENCHMARK_REGISTER_F(NextalignAverageBench, Average)
   ->Unit(benchmark::kMillisecond)//
   ->Complexity(benchmark::oNSquared)
   ->Iterations(20);
 
 
-/**
- * Variation benchmark for nextalign().
- * Runs `nextalign()` for NUM_SEQUENCES_VAR sequences and shows results per sequence.
- * This shows variation or runtime between different sequences.
- */
-void NextalignVariation(benchmark::State& st) {
-  const auto& index = st.range(0);
-  const auto& input = sequences[index];
-  const NextalignOptions options = {};
-  Alignment aln;
-  st.SetLabel(input.seqName);
-  st.SetComplexityN(input.seq.size());
-
-  for (const auto _ : st) {
-    benchmark::DoNotOptimize(aln = nextalign(input.seq, reference, geneMap, options));
-  }
-
-  setCounters(st, 1);
-}
-
-BENCHMARK(NextalignVariation)              //
-  ->DenseRange(0, NUM_SEQUENCES_VAR - 1, 1)//
-  ->Unit(benchmark::kMillisecond)          //
-  ->Complexity(benchmark::oNSquared);
+///**
+// * Variation benchmark for nextalign().
+// * Runs `nextalign()` for NUM_SEQUENCES_VAR sequences and shows results per sequence.
+// * This shows variation or runtime between different sequences.
+// */
+//void NextalignVariation(benchmark::State& st) {
+//  const auto& index = st.range(0);
+//  const auto& input = sequences[index];
+//  const NextalignOptions options = {};
+//  Alignment aln;
+//  st.SetLabel(input.seqName);
+//  st.SetComplexityN(input.seq.size());
+//
+//  for (const auto _ : st) {
+//    benchmark::DoNotOptimize(aln = nextalign(input.seq, reference, geneMap, options));
+//  }
+//
+//  setCounters(st, 1);
+//}
+//
+//BENCHMARK(NextalignVariation)              //
+//  ->DenseRange(0, NUM_SEQUENCES_VAR - 1, 1)//
+//  ->Unit(benchmark::kMillisecond)          //
+//  ->Complexity(benchmark::oNSquared);
