--- packages/nextalign/src/alphabet/letter.h
@@ -1,4 +1,0 @@
-#pragma once
-
-template<typename Letter>
-using Sequence = std::basic_string<Letter>;
